feat: implements JWT and only create, update and delete posts with authenticated and correct user

+++ src/posts/posts.service.ts
@@ -1,7 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, ForbiddenException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Post } from './post.entity';
+import { User } from '../users/user.entity';
 
 @Injectable()
 export class PostsService {
@@ -14,19 +15,29 @@ export class PostsService {
   }
 
   findOne(id: number): Promise<Post> {
-    return this.postsRepository.findOne(id);
+    return this.postsRepository.findOne(id, { relations: ['user'] });
   }
 
-  create(post: Post): Promise<Post> {
+  create(user: User, post: Post): Promise<Post> {
+    post.user = user;
     return this.postsRepository.save(post);
   }
 
-  update(id: number, post: Post): Promise<Post> {
+  update(id: number, user: User, post: Post): Promise<Post> {
     post.id = id;
+
+    if (!user.posts.map(post => post.id).includes(post.id)) {
+      throw new ForbiddenException();
+    }
+
     return this.postsRepository.save(post);
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number, user: User): Promise<void> {
+    if (!user.posts.map(post => post.id).includes(id)) {
+      throw new ForbiddenException();
+    }
+
     await this.postsRepository.delete(id);
   }
 }
